Fixed issue with recursive directories removal. Fixed various memory issues.

+++ Comic/WCComic.h
@@ -14,19 +14,17 @@ typedef enum {
 	__strong ZKDataArchive *archive;
 	Unrar4iOS *rarArchive;
 	NSMutableArray *filesList;
-
 	WCArchType archType;
 }
 
 @property (nonatomic, strong) NSString *file;
 @property (nonatomic, strong) NSString *title;
-@property (nonatomic, readonly) int numberOfPages;
+@property (nonatomic, readonly) NSInteger numberOfPages;
 
 + (void)createCoverImageForFile:(NSString *)path;
 - (id)initWithFile:(NSString *)aFile;
 - (UIImage *)imageAtIndex:(int)index;
 - (void)close;
-
 - (BOOL)somewhereInSubdir:(NSString *)dir;
 
 @end
